add router and fix all pug
