+++ src/App.jsx
@@ -3,20 +3,22 @@ import { BrowserRouter } from 'react-router-dom';
 import{   ShoppingCart,
   ArabicNavbar,
   ProductList,
-  Signin,
-  Signup,
+  SignIn,  
+  SignUp,  
   listofcategory} from'./router'
 
 const App = () => {
   return (
+   
+   
     <BrowserRouter>
-      <div>
-        <Navbar />
-        <div id="vue-root"></div>
-        <sign-in></sign-in> {/* تأكد من تسجيل هذا المكون */}
-        <sign-up></sign-up> {/* تأكد من تسجيل هذا المكون */}
-      </div>
+      <ArabicNavbar />
+      <ProductList />
+      <ShoppingCart />
+      <listofcategory /> {/* Added listofcategory component */}
+      {/* Add other components as needed */}
     </BrowserRouter>
+    
   );
 };
