+++ src/router.js
@@ -1,14 +1,20 @@
-import ShoppingCart from './components/Cart/cart.js';
+// Import React components
+import ShoppingCart from './components/Cart/cart.jsx';
 import ArabicNavbar from './components/Navbar/navbar.jsx';
-import ProductList from './components/Card/productlist.js';
-import Signin from './components/Signin/sign-in.js';
-import Signup from './components/Signup/sign-up.js';
+import ProductList from './components/Card/productlist.jsx';
+
+// Import Lit components
+import { SignIn } from './components/Signin/sign-in.js';
+import { SignUp } from './components/Signup/sign-up.js';
+
+// Import Vue components
 import listofcategory from './components/Categorylist/listofcategory.vue';
+
 export {
     ShoppingCart,
     ArabicNavbar,
     ProductList,
-    Signin,
-    Signup,
+    SignIn, 
+    SignUp,  // Note the curly braces for named exports
     listofcategory
-}
+};
